Added Authentication.cs and created AuthenticationResult and ClaimPrincipalFactory

+++ AuthentificationForm.DataAccess/Repositories/Interfaces/IRoleRepository.cs
@@ -5,6 +5,6 @@ namespace AuthentificationForm.DataAccess.Repositories.Interfaces
     public interface IRoleRepository : IRepository<Role>
     {
         Role GetByName(string name);
-        IList<Role> GetAllByUserId(int id);
+        IList<Role> GetAllByUserId(long id);
     }
 }
